+++ src/components/Query/Query.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import Rules from '../Rules';
+import Rules from './Rules';
 
 const Query = ({ rules, join, codebook, onChange, openDialog }) => (
   <Rules
