+++ src/components/Query/Rules/EditRule.js
@@ -1,10 +1,10 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import Modal from '../../ui/components/Modal';
+import Modal from '@ui/components/Modal';
+import Button from '@ui/components/Button';
 import EditAlterRule from './EditAlterRule';
 import EditEdgeRule from './EditEdgeRule';
 import EditEgoRule from './EditEgoRule';
-import Button from '../../ui/components/Button';
 
 class EditRule extends Component {
   get TypeComponent() {
