+++ src/components/Query/Rules/PreviewRules.js
