+++ src/components/Query/Rules/PreviewText.js
