+++ src/components/Query/Rules/validateRule.js
