+++ src/components/Query/Rules/withAlterRuleType.js
